+++ lld/ELF/InputSection.h
@@ -78,9 +78,6 @@ public:
 
   StringRef name;
 
-  // The 1-indexed partition that this section is assigned to by the garbage
-  // collector, or 0 if this section is dead. Normally there is only one
-  // partition, so this will either be 0 or 1.
   elf::Partition &getPartition(Ctx &) const;
 
   // These corresponds to the fields in Elf_Shdr.
@@ -92,6 +89,7 @@ public:
   uint32_t entsize;
 
   Kind sectionKind;
+  // 0 (dead) or 1 (live).
   uint8_t partition = 1;
 
   // The next two bit fields are only used by InputSectionBase, but we