+++ lld/ELF/Relocations.cpp
@@ -1197,7 +1197,8 @@ template <class ELFT> void elf::scanRelocations(Ctx &ctx) {
     }
     auto scanEH = [&] {
       RelocScan scanner(ctx);
-      for (Partition &part : ctx.partitions) {
+      {
+        Partition &part = *ctx.mainPart;
         for (EhInputSection *sec : part.ehFrame->sections)
           scanner.scanEhSection(*sec);
         if (part.armExidx && part.armExidx->isLive())
@@ -1793,15 +1794,6 @@ ThunkSection *ThunkCreator::addThunkSection(OutputSection *os,
   return ts;
 }
 
-static bool isThunkSectionCompatible(InputSection *source,
-                                     SectionBase *target) {
-  // We can't reuse thunks in different loadable partitions because they might
-  // not be loaded. But partition 1 (the main partition) will always be loaded.
-  if (source->partition != target->partition)
-    return target->partition == 1;
-  return true;
-}
-
 std::pair<Thunk *, bool> ThunkCreator::getThunk(InputSection *isec,
                                                 Relocation &rel, uint64_t src) {
   SmallVector<std::unique_ptr<Thunk>, 0> *thunkVec = nullptr;
@@ -1826,8 +1818,7 @@ std::pair<Thunk *, bool> ThunkCreator::getThunk(InputSection *isec,
 
   // Check existing Thunks for Sym to see if they can be reused
   for (auto &t : *thunkVec)
-    if (isThunkSectionCompatible(isec, t->getThunkTargetSym()->section) &&
-        t->isCompatibleWith(*isec, rel) &&
+    if (t->isCompatibleWith(*isec, rel) &&
         ctx.target->inBranchRange(rel.type, src,
                                   t->getThunkTargetSym()->getVA(ctx, -pcBias)))
       return std::make_pair(t.get(), false);